+++ core/src/main/kotlin/com/robohorse/robopojogenerator/models/GenerationModel.kt
@@ -18,7 +18,9 @@ data class GenerationModel(
     val kotlinNullableFields: Boolean,
     val javaPrimitives: Boolean,
     val useTabsIndentation: Boolean,
-    val useLombokValue: Boolean
+    val useLombokValue: Boolean,
+    val useMoshiAdapter: Boolean,
+    val useKotlinDataClass: Boolean
 )
 
 data class ProjectModel(