+++ generator/src/main/kotlin/com/robohorse/robopojogenerator/postrocessing/common/KotlinDataClassPostProcessor.kt
@@ -8,6 +8,7 @@ import com.robohorse.robopojogenerator.models.FrameworkVW.LoganSquare
 import com.robohorse.robopojogenerator.models.FrameworkVW.Moshi
 import com.robohorse.robopojogenerator.models.GenerationModel
 import com.robohorse.robopojogenerator.postrocessing.BasePostProcessor
+import com.robohorse.robopojogenerator.postrocessing.utils.MoshiAnnotationsProcessor
 import com.robohorse.robopojogenerator.properties.ClassItem
 import com.robohorse.robopojogenerator.properties.annotations.KotlinAnnotations
 import com.robohorse.robopojogenerator.properties.templates.ClassTemplate
@@ -19,7 +20,8 @@ import com.robohorse.robopojogenerator.utils.ClassTemplateHelper
 
 internal class KotlinDataClassPostProcessor(
     generateHelper: ClassGenerateHelper,
-    classTemplateHelper: ClassTemplateHelper
+    classTemplateHelper: ClassTemplateHelper,
+    private val moshiAnnotationsProcessor: MoshiAnnotationsProcessor
 ) : BasePostProcessor(generateHelper, classTemplateHelper) {
 
     override fun proceedClassImports(
@@ -85,6 +87,7 @@ internal class KotlinDataClassPostProcessor(
                 ImportsTemplate.GSON.imports
             )
         }
+
         is LoganSquare -> {
             generateHelper.setAnnotations(
                 classItem,
@@ -93,6 +96,7 @@ internal class KotlinDataClassPostProcessor(
                 ImportsTemplate.LOGAN_SQUARE.imports
             )
         }
+
         is Jackson -> {
             generateHelper.setAnnotations(
                 classItem,
@@ -101,6 +105,7 @@ internal class KotlinDataClassPostProcessor(
                 ImportsTemplate.JACKSON.imports
             )
         }
+
         is FastJson -> {
             generateHelper.setAnnotations(
                 classItem,
@@ -109,14 +114,9 @@ internal class KotlinDataClassPostProcessor(
                 ImportsTemplate.FAST_JSON.imports
             )
         }
-        is Moshi -> {
-            generateHelper.setAnnotations(
-                classItem,
-                KotlinAnnotations.MOSHI.classAnnotation,
-                KotlinAnnotations.MOSHI.annotation,
-                ImportsTemplate.MOSHI.imports
-            )
-        }
+
+        is Moshi -> moshiAnnotationsProcessor.applyAnnotations(generationModel, classItem)
+
         else -> {
             // NO OP
         }
@@ -129,6 +129,6 @@ internal class KotlinDataClassPostProcessor(
     ) = classTemplateHelper.createClassBodyKotlinDataClass(
         classItem,
         classBody,
-        generationModel.useKotlinParcelable
+        generationModel
     )
 }